back/requirements.txt
Django==4.2.7
fastapi==0.104.1
uvicorn==0.24.0
asgiref==3.7.2
sqlparse==0.4.4
typing-extensions==4.8.0
pydantic==2.5.2
starlette==0.27.0
databases==0.6.0 
psycopg2-binary==2.9.6
asyncpg==0.27.0
scapy==2.5.0
pyroute2==0.7.9
netaddr==0.8.0
requests==2.31.0
httpx==0.25.1
python-dotenv==1.0.0
pyyaml==6.0.1
netmiko==4.2.0
paramiko==3.3.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-django==4.7.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
tzdata==2024.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
python-jose==3.3.0
python-multipart==0.0.6